backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
faiss-cpu
sentence-transformers
numpy
aiofiles
